stable-requirements.txt
astroid==2.2.5
Click==7.0
Flask==1.0.3
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.16
numpy==1.16.4
protobuf==3.8.0
pylint==2.3.1
python-dateutil==2.8.0
python-slugify==3.0.2
six==1.12.0
text-unidecode==1.2
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.1
wwdtm==0.8.0
